Add basic format check
Prior to this commit, there was no checking to ensure that a passed value was a
4 segment string. This change adds such checking, and ensures NaN is returned
where appropriate.

The reasoning for returning NaN is thus: we're converting to a number, so we're
expecting to do some sort of comparison with it - which will always fail in the
case of NaN being compared. Secondly, an IP is a representation of a number - if
the format is incorrect then it is technically not a number and thus NaN is
correct.

The reasoning for not checking the validity of an IP is mostly practical - this
module is a single-concern module. Checking the validity of an IP should be done
in a separate module, as that requires it be explicitly included (sometimes, we
may either trust the original source of the data, or simply not care).